requirements.txt
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
faiss-gpu==1.7.2
filelock==3.8.0
huggingface-hub==0.10.0
idna==3.4
joblib==1.2.0
nltk==3.7
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
soupsieve==2.3.2.post1
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tqdm==4.64.1
transformers==4.22.2
typing_extensions==4.4.0
urllib3==1.26.12
